+++ SHM/src/main/java/ua/com/manometer/service/modeldescription/ModelDescriptionService.java
@@ -12,10 +12,10 @@ public interface ModelDescriptionService {
 
 	public List<ModelDescription> listModelDescription();
 
-	public void removeModelDescription(Long id);
+	public void removeModelDescription(Integer id);
 
-    public  List<ModelDescription> findListByIds(List<Long> modelIds);
+    public  List<ModelDescription> findListByIds(List<Integer> modelIds);
 
 
-    public ModelDescription getModelDescription(Long model);
+    public ModelDescription getModelDescription(Integer model);
 }
